Add check for web3 status
